Expanded the class to include security based list functions.

+++ dotproject/classes/dp.class.php
@@ -1,17 +1,17 @@
 <?php /* CLASSES $Id$ */
 
 ##
-## CdpObject Abstract Class
+## CDpObject Abstract Class
 ##
 
-class CdpObject {
+class CDpObject {
 // table name
 	var $_tbl = '';
 // table primary key field
 	var $_tbl_key = '';
 
 // object constructor to set table and key field
-	function CdpObject( $table, $key ) {
+	function CDpObject( $table, $key ) {
 		$this->_tbl = $table;
 		$this->_tbl_key = $key;
 	}
@@ -51,8 +51,21 @@ class CdpObject {
 		}
 	}
 
+// default check for delete dependancies (can be overloaded)
+	function canDelete( &$msg, $oid=null ) {
+		if ($oid) {
+			$k = $this->_tbl_key;
+			$this->$k = intval( $oid );
+		}
+		return true;
+	}
+
 // default delete method (can be overloaded)
 	function delete() {
+		if (!$this->canDelete( $msg )) {
+			return $msg;
+		}
+
 		$k = $this->_tbl_key;
 		$sql = "DELETE FROM $this->_tbl WHERE $this->_tbl_key = '".$this->$k."'";
 		if (!db_exec( $sql )) {
@@ -60,6 +73,43 @@ class CdpObject {
 		} else {
 			return NULL;
 		}
+	}
+
+// get specifically denied records from a module based on a user
+	function getDeniedRecords( $uid ) {
+		$uid || exit ("FATAL ERROR<br />" . get_class( $this ) . "::getDeniedRecords failed" );
+
+		// get read denied projects
+		$deny = array();
+		$sql = "
+		SELECT $this->_tbl_key
+		FROM $this->_tbl, permissions
+		WHERE permission_user = $uid
+			AND permission_grant_on = '$this->_tbl'
+			AND permission_item = $this->_tbl_key
+			AND permission_value = 0
+		";
+		return db_loadColumn( $sql );
+	}
+
+// returns a list of records exposed to the user
+	function getAllowedRecords( $uid, $fields='*', $orderby='', $index=null ) {
+		$uid || exit ("FATAL ERROR<br />" . get_class( $this ) . "::getAllowedRecords failed" );
+		$deny = $this->getDeniedRecords( $uid );
+
+		$sql = "SELECT $fields"
+			. "\nFROM $this->_tbl, permissions"
+			. "\nWHERE permission_user = $uid"
+			. "\n	AND permission_value <> 0"
+			. "\n	AND ("
+			. "\n		(permission_grant_on = 'all')"
+			. "\n		OR (permission_grant_on = '$this->_tbl' AND permission_item = -1)"
+			. "\n		OR (permission_grant_on = '$this->_tbl' AND permission_item = $this->_tbl_key)"
+			. "\n	)"
+			. (count($deny) > 0 ? "\nAND $this->_tbl_key NOT IN (" . implode( ',', $deny ) . ')' : '')
+			. ($orderby ? "\nORDER BY $orderby" : '');
+
+		return db_loadHashList( $sql, $index );	
 	}
 }
 ?>
